requirements.txt
torch==1.3.1
torchvision
numpy
tqdm
Pillow
protobuf
tensorboardX